root/python/services/message-service/message-service-grpc/requirements.txt
grpcio
grpcio-tools
protobuf
prometheus-client
